Refactored safety.py into compat.py
